requirements.txt
fastapi
numpy
pandas
uvicorn
pydantic
scikit-learn
facebook-business
python-dotenv
matplotlib
plotly
jinja2